updated src and makefile

+++ src/mle_sumcheck.cpp
@@ -56,7 +56,7 @@ bool sVerifier::execute_sumcheck(sProver& pr, const ligeropcs_base& oracle, cons
 
     // s_{i - 1}
     std::array<Goldilocks2::Element, 2> si1;
-    for(int round = 1;round <= nrnd; ++round){
+    for(size_t round = 1;round <= nrnd; ++round){
         // s_i
         std::array<Goldilocks2::Element, 2> si;    
         si = pr.send_message(round, challenges);
@@ -113,7 +113,7 @@ bool sVerifier::execute_sumcheck(sProver& pr, const ligeropcs_ext& oracle, const
 
     // s_{i - 1}
     std::array<Goldilocks2::Element, 2> si1;
-    for(int round = 1;round <= nrnd; ++round){
+    for(size_t round = 1;round <= nrnd; ++round){
         // s_i
         std::array<Goldilocks2::Element, 2> si;    
         si = pr.send_message(round, challenges);
